+++ src/Faker/Internal/Types.hs
@@ -22,6 +22,7 @@ data SourceData
   | Bank
   | Beer
   | BoJackHorseman
+  | BigBangTheory
   | Book
   | Blood
   | Chiquito